# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing products with JWT authentication",
        "title": "Product API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Lokesh Choraria",
            "email": "lokeshchoraria2@email.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/login": {
            "post": {
                "description": "Authenticates a registered user using username and password. Returns a JWT token upon success.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authentication successful – JWT access token returned",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request – Invalid JSON or missing fields",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized – Incorrect password",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found – User does not exist",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/products/extreme": {
            "get": {
                "description": "Fetches either the product with the highest or lowest quantity based on query parameter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get product with extreme quantity",
                "parameters": [
                    {
                        "type": "boolean",
                        "example": true,
                        "description": "Set to true to get product with highest quantity",
                        "name": "most",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "example": false,
                        "description": "Set to true to get product with lowest quantity",
                        "name": "least",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    "400": {
                        "description": "Missing or conflicting query parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/products/get": {
            "get": {
                "description": "Retrieves a single product based on the provided UUID in query parameter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get a product by ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"d290f1ee-6c54-4b01-90e6-d701748f0851\"",
                        "description": "Product UUID",
                        "name": "product_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    "400": {
                        "description": "Missing or invalid product_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "description": "Creates a new user in the system with a unique username and email. The password is securely hashed before storage.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration payload",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User successfully registered",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request – Invalid JSON or missing fields",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict – Username or email already exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "strongPassword123"
                },
                "username": {
                    "type": "string",
                    "example": "john_doe"
                }
            }
        },
        "models.Product": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-07-25T14:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "A bright red cotton t-shirt"
                },
                "id": {
                    "type": "string",
                    "example": "2c8a21e3-c882-4b40-9f27-35413e5e64e7"
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/images/redshirt.png"
                },
                "name": {
                    "type": "string",
                    "example": "Red T-Shirt"
                },
                "price": {
                    "type": "number",
                    "example": 19.99
                },
                "quantity": {
                    "type": "integer",
                    "example": 42
                },
                "sku": {
                    "type": "string",
                    "example": "RTS-XL-001"
                },
                "type": {
                    "type": "string",
                    "example": "Clothing"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-07-25T14:30:00Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "bfc5b2b1-bc0e-4f2b-8c18-7c7a47fdc9c4"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-07-25T14:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "strongPassword123"
                },
                "user_id": {
                    "type": "string",
                    "example": "bfc5b2b1-bc0e-4f2b-8c18-7c7a47fdc9c4"
                },
                "username": {
                    "type": "string",
                    "example": "john_doe"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}